example/pyproject.toml
[project]
name = "fastapi-sendparcel-example"
version = "0.1.0"
description = "Example app demonstrating fastapi-sendparcel with Tabler + HTMX"
requires-python = ">=3.12"
dependencies = [
    "fastapi-sendparcel",
    "fastapi>=0.115.0",
    "jinja2>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.6",
]

[tool.uv.sources]
fastapi-sendparcel = { path = "..", editable = true }
